+++ packages/adapters/uni-uploadFile/index.ts
@@ -2,12 +2,12 @@ import { Adapter } from '@applet-request/core';
 import type { RequestContext, MiddlewareNext, RequestConfig } from '@applet-request/core';
 
 /**
- * uni.downloadFile的其他请求配置
+ * uni.uploadFile的其他请求配置
  */
-export type UniUploadFileOtherConfig = Omit<UniApp.UploadFileOption, 'url' | 'fail' | 'success' | 'complete'>;
+export type UniUploadFileOtherConfig = Omit<UniApp.UploadFileOption, 'url' | 'fail' | 'success' | 'complete' | 'formData'>;
 
 /**
- * uni.downloadFile的请求配置
+ * uni.uploadFile的请求配置
  */
 export type UniUploadFileConfig = RequestConfig<UniUploadFileOtherConfig>;
 
@@ -32,6 +32,7 @@ export class UniUploadFileAdaptor<Data> extends Adapter<
       uni.uploadFile({
         ...context.request.config,
         url: context.request.apiURL,
+        formData: context.request.data,
         success(res) {
           context.response.data = res.data as Data;
           context.response.raw = res;